poc/rag/poc_dev_env_20250714/pyproject.toml
[project]
name = "poc-dev-env-20250714"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.57.1",
    "google-cloud-dlp>=3.31.0",
    "numpy>=2.2.6",
    "openai>=1.95.1",
    "pinecone>=7.3.0",
]
